+++ CotEditor/Sources/NSString+Normalization.m
@@ -85,4 +85,21 @@
     return result;
 }
 
+
+// ------------------------------------------------------
+/// A string made by normalizing the receiver’s contents using the normalization form adopted by HFS+, a.k.a. Apple Modified NFD
+- (nonnull NSString *)decomposedStringWithHFSPlusMapping
+// ------------------------------------------------------
+{
+    CFStringRef sourceStr = (__bridge CFStringRef)self;
+    CFIndex length = CFStringGetMaximumSizeOfFileSystemRepresentation(sourceStr);
+    char *destStr = (char *)malloc(length);
+    
+    Boolean success = CFStringGetFileSystemRepresentation(sourceStr, destStr, length);
+    NSString *result = success ? [NSString stringWithUTF8String:destStr] : self;
+    free(destStr);
+    
+    return result;
+}
+
 @end